pyproject.toml
[project]
name = "uzum"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "boto3>=1.42.46",
    "celery>=5.6.2",
    "celery-types>=0.24.0",
    "django>=6.0.2",
    "django-celery-beat>=2.1.0",
    "django-ckeditor-5>=0.2.19",
    "django-elasticsearch-dsl-drf>=0.22.5",
    "django-filter>=25.2",
    "django-jsonform>=2.23.2",
    "django-location-field>=2.7.3",
    "django-minio-storage>=0.5.9",
    "django-mptt>=0.18.0",
    "django-storages>=1.14.6",
    "django-unfold>=0.80.2",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "elasticsearch>=9.3.0",
    "image>=1.5.33",
    "moviepy==1.0.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
]
